test: move system call tests
