+++ crates/evm2/src/evm/mod.rs
@@ -834,119 +834,6 @@ mod tests {
     }
 
     #[test]
-    fn system_call_uses_system_sender_without_fee_accounting() {
-        let contract = Address::from([0x42; 20]);
-        let beneficiary = Address::from([0x99; 20]);
-        let code = Bytecode::new_legacy(Bytes::from_static(&[
-            op::CALLER,
-            op::PUSH1,
-            0,
-            op::SSTORE,
-            op::ORIGIN,
-            op::PUSH1,
-            1,
-            op::SSTORE,
-            op::STOP,
-        ]));
-        let mut database = InMemoryDB::default();
-        database.insert_account_info(contract, AccountInfo::default().with_code(code));
-        let block = BlockEnv { beneficiary, basefee: U256::from(7), ..BlockEnv::default() };
-        let mut evm = Evm::<TestEvmTypes>::new(
-            SpecId::OSAKA,
-            block,
-            TxRegistry::new(),
-            database,
-            Precompiles::base(SpecId::OSAKA),
-        );
-
-        let result = evm.system_call(contract, Bytes::new());
-
-        assert!(result.status);
-        assert!(result.gas_used < SYSTEM_CALL_GAS_LIMIT);
-        assert!(!result.state_changes.accounts.contains_key(&SYSTEM_ADDRESS));
-        assert!(!result.state_changes.accounts.contains_key(&beneficiary));
-        let storage = result.state_changes.storage.get(&contract).expect("storage changed");
-        let system_address = U256::from_be_slice(SYSTEM_ADDRESS.as_slice());
-        assert_eq!(storage.slots.get(&U256::ZERO).map(|slot| slot.current), Some(system_address));
-        assert_eq!(storage.slots.get(&U256::ONE).map(|slot| slot.current), Some(system_address));
-    }
-
-    #[test]
-    fn system_call_starts_with_warm_system_contract() {
-        let contract = Address::from([0x42; 20]);
-        let code =
-            Bytecode::new_legacy(Bytes::from_static(&[op::ADDRESS, op::EXTCODESIZE, op::STOP]));
-        let mut database = InMemoryDB::default();
-        database.insert_account_info(contract, AccountInfo::default().with_code(code));
-        let mut evm = Evm::<TestEvmTypes>::new(
-            SpecId::OSAKA,
-            BlockEnv::default(),
-            TxRegistry::new(),
-            database,
-            Precompiles::base(SpecId::OSAKA),
-        );
-
-        let result = evm.system_call(contract, Bytes::new());
-
-        assert!(result.status);
-        assert!(
-            result.gas_used < 1_000,
-            "system contract should be warm before execution, got {} gas used",
-            result.gas_used
-        );
-    }
-
-    #[test]
-    fn system_call_to_missing_code_is_noop() {
-        let contract = Address::from([0x42; 20]);
-        let mut evm = Evm::<TestEvmTypes>::new(
-            SpecId::OSAKA,
-            BlockEnv::default(),
-            TxRegistry::new(),
-            InMemoryDB::default(),
-            Precompiles::base(SpecId::OSAKA),
-        );
-
-        let result = evm.system_call(contract, Bytes::new());
-
-        assert!(result.status);
-        assert_eq!(result.gas_used, 0);
-        assert!(result.state_changes.is_empty());
-    }
-
-    #[test]
-    fn system_call_reverts_state_changes() {
-        let contract = Address::from([0x42; 20]);
-        let code = Bytecode::new_legacy(Bytes::from_static(&[
-            op::PUSH1,
-            1,
-            op::PUSH1,
-            0,
-            op::SSTORE,
-            op::PUSH1,
-            0,
-            op::PUSH1,
-            0,
-            op::REVERT,
-        ]));
-        let mut database = InMemoryDB::default();
-        database.insert_account_info(contract, AccountInfo::default().with_code(code));
-        let mut evm = Evm::<TestEvmTypes>::new(
-            SpecId::OSAKA,
-            BlockEnv::default(),
-            TxRegistry::new(),
-            database,
-            Precompiles::base(SpecId::OSAKA),
-        );
-
-        let result = evm.system_call(contract, Bytes::new());
-
-        assert!(!result.status);
-        assert_eq!(result.stop, InstrStop::Revert);
-        assert!(result.state_changes.is_empty());
-    }
-
-    #[test]
     fn host_executes_message() {
         let mut evm = Evm::<TestEvmTypes>::new(
             SpecId::OSAKA,
